requirements.txt
setuptools==69.0.2
python-docx==1.1.0
pandas==2.0.3
Pillow==10.1.0
requests==2.31.0
Flask==3.0.0
